rfnoc: fft: Fix scaling factor property type
The property `fft_scaling` in the new FFT block was defined as a float,
but should be a double. Rationale:

- The corresponding public APIs already use `double`, e.g.,
  set_scaling_factor(), already used doubles. That means we have
  2 equivalent APIs (property API and regular C++ API) with competing
    types.
- `float` is potentially not accurate enough to store the actual,
  desired scaling factor.
- Most importantly, this breaks the ability to set the property from
  Python, e.g., when in GNU Radio. That's because in Python, there's
  only one real data type, and if we see that, we try and convert it to
  `double`.

With this fix, it's now possible to run the FFT block from GNU Radio
using a generic RFNoC block controller.

+++ host/lib/rfnoc/fft_block_control.cpp
@@ -156,7 +156,7 @@ public:
 
     void set_scaling_factor(const double factor) override
     {
-        set_property<float>(PROP_KEY_FFT_SCALING_FACTOR, static_cast<double>(factor));
+        set_property<double>(PROP_KEY_FFT_SCALING_FACTOR, static_cast<double>(factor));
     }
 
     void set_scaling(const uint32_t scaling) override
@@ -739,7 +739,7 @@ private:
         PROP_KEY_DIRECTION, static_cast<int>(DEFAULT_DIRECTION), {res_source_info::USER}};
     property_t<int> _scaling = property_t<int>{
         PROP_KEY_FFT_SCALING, DEFAULT_FFT_SCALING, {res_source_info::USER}};
-    property_t<float> _scaling_factor = property_t<float>{
+    property_t<double> _scaling_factor = property_t<double>{
         PROP_KEY_FFT_SCALING_FACTOR, DEFAULT_FFT_SCALING_FACTOR, {res_source_info::USER}};
     property_t<int> _shift = property_t<int>{
         PROP_KEY_SHIFT_CONFIG, static_cast<int>(DEFAULT_SHIFT), {res_source_info::USER}};
